requirements.txt
Flask>=2.0.0
requests>=2.25.0
python-dotenv>=0.15.0
colorthief>=0.2.1